chore: extract core/game_log, core/game_strings from utils/game_utils
