+++ configure.ac
@@ -1,5 +1,5 @@
 # Define the package name and version
-AC_INIT([helapordo], [1.4.19], [jgabaut@github.com])
+AC_INIT([helapordo], [1.4.20], [jgabaut@github.com])
 
 # Verify automake version and enable foreign option
 AM_INIT_AUTOMAKE([foreign -Wall])
@@ -193,7 +193,7 @@ AM_CONDITIONAL([USE_DEFAULT_BACKGROUND], [test "$enable_default_background" = "y
 # Set a default version number if not specified externally
 AC_ARG_VAR([VERSION], [Version number])
 if test -z "$VERSION"; then
-  VERSION="1.4.19"
+  VERSION="1.4.20"
 fi
 
 # Output variables to the config.h header
