+++ kazoj/bone/const_rl.c
@@ -12,7 +12,7 @@ void fail(char* msg, int ex, int val) {
 
 int main(void) {
  int check = -1;
- if ( ! (( check = strcmp(VERSION, "1.4.19") ) == 0)) {
+ if ( ! (( check = strcmp(VERSION, "1.4.20") ) == 0)) {
    fail("VERSION",0,check);
  };
  if ( ! (HLP_MAX_INDEX == 35) ) {
