+++ src/utils/game_utils.c
@@ -177,31 +177,6 @@ void printWin_EnvVars(void)
     printf("  Number_Of_Processors: { %s }\n", getenv("Number_Of_Processors"));
     printf("  OS: { %s }\n", getenv("OS"));
 }
-
-/**
- * Logs Windows envvars to debug log file.
- */
-void log_Win_EnvVars(void)
-{
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "Windows Environment vars:");
-    log_tag("debug_log.txt", "[WIND32-DEBUG", "UserProfile: { %s }",
-            getenv("UserProfile"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "HomeDrive: { %s }",
-            getenv("HomeDrive"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "HomePath: { %s }",
-            getenv("HomePath"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "ComputerName: { %s }",
-            getenv("ComputerName"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "Processor_Revision: { %s }",
-            getenv("Processor_Revision"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "Processor_Identifier: { %s }",
-            getenv("Processor_Identifier"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "Processor_Level: { %s }",
-            getenv("Processor_Level"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "Number_Of_Processors: { %s }",
-            getenv("Number_Of_Processors"));
-    log_tag("debug_log.txt", "[WIN32-DEBUG]", "OS: { %s }", getenv("OS"));
-}
 #endif
 
 void dbg_GameOptions(GameOptions * options)
@@ -717,18 +692,6 @@ saveType saveTypeFrom_string(char *s)
 }
 
 /**
- * Logs a turnOption_OP to "$STATICDIR/OPS_log.txt".
- * @param op The turnOption_OP to log.
- */
-void log_OP(turnOption_OP op)
-{
-    log_tag(OPS_LOGFILE, "[LOG]", "[OP_code: %i, value: %s]", op,
-            stringFromTurnOP(op));
-    log_tag("debug_log.txt", "[LOG_TURNOP]", "[OP_code: %i, value: %s]", op,
-            stringFromTurnOP(op));
-}
-
-/**
  * Prints an ASCII escape code that makes subsequent output a thick red color.
  */
 void red(void)
@@ -871,73 +834,6 @@ void init_game_color_pairs(void)
 }
 #endif // HELAPORDO_RAYLIB_BUILD
 #endif // HELAPORDO_CURSES_BUILD
-
-/**
- * Sets the passed char array to the expected path for /static/ folder.
- * @param static_path The array that will hold path to static folder.
- */
-void resolve_staticPath(char static_path[500])
-{
-    char homedir_path[200];
-#ifndef _WIN32
-    sprintf(homedir_path, "%s", getenv("HOME"));
-#else
-    sprintf(homedir_path, "%s", getenv("UserProfile"));
-#endif
-
-#ifndef _WIN32
-    const char *static_folder_path_wd = "./";
-#else
-    const char *static_folder_path_wd = ".\\";
-#endif
-
-#ifndef _WIN32
-    const char *local_install_static_folder_path = "/helapordo-local/";
-#else
-    const char *local_install_static_folder_path = "\\helapordo-local\\";
-#endif
-    char static_folder_path_global[500];
-    sprintf(static_folder_path_global, "%s", homedir_path);
-    strncat(static_folder_path_global, local_install_static_folder_path, 50);
-    struct stat sb;
-
-    if (G_USE_CURRENTDIR == 1 ) {
-        if (stat(static_folder_path_wd, &sb) == 0 && S_ISDIR(sb.st_mode)) {
-            //sprintf(msg, "[DEBUG]    resolve_staticPath(): Found \"/static/\" dir in working directory (%s).\n",static_folder_path_wd);
-            strcpy(static_path, static_folder_path_wd);
-        } else {
-            fprintf(stderr, "\n[ERROR]    Can't find static dir. Quitting.\n");
-            fprintf(stderr, "\nWorking static dir at: (%s).\n",
-                    static_folder_path_wd);
-            exit(EXIT_FAILURE);
-        }
-    } else {
-        //sprintf(msg, "[DEBUG]    resolve_staticPath(): Can't find \"/static/\" dir in working directory (%s). Will try \"%s/helapordo-local/static/\".\n", static_folder_path_wd, homedir_path);
-        if (stat(static_folder_path_global, &sb) == 0 && S_ISDIR(sb.st_mode)) {
-            //sprintf(msg, "[DEBUG]    resolve_staticPath(): Found \"/static/\" dir in global directory: \"%s/helapordo-local/static/\".\n", homedir_path);
-            strcpy(static_path, static_folder_path_global);
-        } else {
-#ifndef _WIN32
-            int mkdir_global_res = mkdir(static_folder_path_global, 0777);
-#else
-            int mkdir_global_res = mkdir(static_folder_path_global);
-#endif
-            if (mkdir_global_res != 0) {
-                //sprintf(msg,"[DEBUG]    resolve_staticPath(): Can't find \"/static/\" dir in \"%s/helapordo-local/static/\". Quitting.\n", homedir_path);
-                fprintf(stderr, "\n[ERROR]    Can't find static dir. Quitting.\n");
-                fprintf(stderr, "\nHome dir at: (%s).\n", homedir_path);
-                fprintf(stderr, "\nGlobal static dir at: (%s).\n",
-                        static_folder_path_global);
-                fprintf(stderr, "\nWorking static dir at: (%s).\n",
-                        static_folder_path_wd);
-                exit(EXIT_FAILURE);
-            } else {
-                fprintf(stderr, "%s():    Could not find {%s} at first, so it was created.\n", __func__, static_folder_path_global);
-                strcpy(static_path, static_folder_path_global);
-            }
-        }
-    }
-}
 
 /**
  * Loads the lore strings from the appropriate lore file.
@@ -1301,293 +1197,6 @@ void usage(const char *progname)
 }
 
 /**
- * Takes a filename string, a string headear and a format string.
- * Tries logging the message to the passed file if global var G_DEBUG_ON is set.
- * @param filename The filename to open.
- * @param header The string header for the message to log.
- * @param format The format string for message.
- */
-void log_tag(char *filename, char *header, const char *format, ...)
-{
-#ifndef HELAPORDO_DEBUG_LOG
-#else
-    // Open log file if log flag is set and append to it
-    if (G_LOG_ON == 1) {
-        char path_to_debug_file[500];
-        char static_path[500];
-
-        // Set static_path value to the correct static dir path
-        resolve_staticPath(static_path);
-
-        sprintf(path_to_debug_file, "%s", static_path);
-
-#ifndef _WIN32
-        strncat(path_to_debug_file, "/", 2);
-#else
-        strncat(path_to_debug_file, "\\", 2);
-#endif
-
-        strncat(path_to_debug_file, filename, 200);
-
-        //fprintf(stderr, "Using %s as path to debug log.\n", path_to_debug_file);
-
-        FILE *logfile = fopen(path_to_debug_file, "a");
-        if (!logfile) {
-            fprintf(stderr,
-                    "Error opening log file.\n Static path: (%s) Filename : (%s).\n",
-                    static_path, filename);
-            fprintf(stderr, "Path to debug file was: (%s).",
-                    path_to_debug_file);
-            fprintf(stderr, "Format was:    %s", format);
-            exit(EXIT_FAILURE);
-        }
-        va_list args;
-        va_start(args, format);
-        time_t now = time(0);
-        struct tm *mytime = localtime(&now);
-        char timeheader[500];
-        if (strftime(timeheader, sizeof timeheader, "%X", mytime)) {
-            fprintf(logfile, "[ %s ] [ %-12.12s ] [", timeheader, header);
-            vfprintf(logfile, format, args);
-            fprintf(logfile, "]\n");
-        }
-        va_end(args);
-        fclose(logfile);
-    }
-#endif
-}
-
-/**
- * Takes a OP_res and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see OP_res
- * @param r The OP_res.
- * @return String corresponding to the OP result.
- */
-char *stringFrom_OP_res(OP_res r)
-{
-
-    return opresstrings[r];
-}
-
-/**
- * Takes a saveType and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see saveType
- * @param s The saveType value.
- * @return String corresponding to the save type.
- */
-char *stringFrom_saveType(saveType s)
-{
-
-    return savetypestrings[s];
-}
-
-/**
- * Takes a turnOption_OP and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see turnOption_OP
- * @param t The turn choice.
- * @return String corresponding to the turn choice.
- */
-char *stringFromTurnOP(turnOption_OP t)
-{
-
-    return turnopstrings[t];
-}
-
-/**
- * Takes a foeTurnOption_OP and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see foeTurnOption_OP
- * @param ft The turn choice.
- * @return String corresponding to the turn choice.
- */
-char *stringFromFoeTurnOP(foeTurnOption_OP ft)
-{
-    return foeturnopstrings[ft];
-}
-
-/**
- * Takes a fightResult fr and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see fightResult
- * @param fr The fight result value.
- * @return String corresponding to the fight result value.
- */
-char *stringFrom_fightResult(fightResult fr)
-{
-
-    return fightresultstrings[fr];
-}
-
-/**
- * Takes a fighterStatus and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see fighterStatus
- * @param s The status.
- * @return String corresponding to the status.
- */
-char *stringFromStatus(fighterStatus s)
-{
-    static char *strings[] = {
-        "Normal",
-        "Poison",
-        "Frozen",
-        "Burned",
-        "Weak",
-        "Strong"
-    };
-
-    return strings[s];
-}
-
-/**
- * Takes a winconClass and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see winconClass
- * @param w The class.
- * @return String corresponding to the class.
- */
-char *stringFromWinconClass(winconClass w)
-{
-
-    return winconstrings[w];
-}
-
-/**
- * Takes a fighterClass and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see fighterClass
- * @param f The class.
- * @return String corresponding to the class.
- */
-char *stringFromClass(fighterClass f)
-{
-
-    return classesstrings[f];
-}
-
-/**
- * Takes a enemyClass and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see enemyClass
- * @param e The class.
- * @return String corresponding to the class.
- */
-char *stringFromEClass(enemyClass e)
-{
-
-    return classenemystrings[e];
-}
-
-/**
- * Takes a bossClass and returns the corresponding string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see bossClass
- * @param b The class.
- * @return String corresponding to the class.
- */
-char *stringFromBossClass(bossClass b)
-{
-
-    return classbossstrings[b];
-}
-
-/**
- * Takes a integer and returns the corresponding consumable string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see consumableClass
- * @param c The integer/consumableClass.
- * @return String corresponding to the consumable.
- */
-char *stringFromConsumables(int c)
-{
-
-    return consumablestrings[c];
-}
-
-/**
- * Takes a integer and returns the corresponding equip string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see equipClass
- * @param e The integer/equipClass.
- * @return String corresponding to the equip.
- */
-char *stringFromEquips(int e)
-{
-
-    return equipstrings[e];
-}
-
-/**
- * Takes a integer and returns the corresponding equipzone string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see Equipzone
- * @param z The integer/Equipzone.
- * @return String corresponding to the equipzone.
- */
-char *stringFromEquipzones(int z)
-{
-
-    return equipzonestrings[z];
-}
-
-/**
- * Takes a integer and returns the corresponding quality string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see quality
- * @param q The integer/quality
- * @return String corresponding to the quality.
- */
-char *stringFromQuality(int q)
-{
-
-    return qualitytrings[q];
-}
-
-/**
- * Takes a integer and returns the corresponding artifact string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see artifactClass
- * @param a The integer/artifactClass.
- * @return String corresponding to the artifact.
- */
-char *stringFromArtifacts(int a)
-{
-
-    return artifactstrings[a];
-}
-
-/**
- * Takes a fighterClass and a specialMove and returns the corresponding name string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see fighterClass
- * @see specialMove
- * @param class The fighterClass.
- * @param move The specialMove.
- * @return String corresponding to the special move.
- */
-char *nameStringFromSpecial(fighterClass class, specialMove move)
-{
-    return specialsnamestrings[class][(move % (SPECIALSMAX + 1))];
-}
-
-/**
- * Takes a fighterClass and a specialMove and returns the corresponding desc string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see fighterClass
- * @see specialMove
- * @param class The fighterClass.
- * @param move The specialMove.
- * @return String corresponding to the special move.
- */
-char *descStringFromSpecial(fighterClass class, specialMove move)
-{
-    return specialsdescstrings[class][(move % (SPECIALSMAX + 1))];
-}
-
-/**
  * Takes a fighterClass and a specialMove and returns the corresponding cost by the inner array position, as an integer.
  * Correct result is only possible by having the enum values in a consistent order with the integer array.
  * @see fighterClass
@@ -1602,32 +1211,6 @@ int costFromSpecial(fighterClass class, specialMove move)
 }
 
 /**
- * Takes a skillType and returns the corresponding name string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see fighterClass
- * @see specialMove
- * @param class The skillType.
- * @return String corresponding to the skillType.
- */
-char *nameStringFromSkill(skillType class)
-{
-    return skillsnamestrings[class];
-}
-
-/**
- * Takes a fighterClass and a specialMove and returns the corresponding desc string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see fighterClass
- * @see skillType
- * @param class The skillType.
- * @return String corresponding to the skillType.
- */
-char *descStringFromSkill(skillType class)
-{
-    return skillsdescstrings[class];
-}
-
-/**
  * Takes a skillType and returns the corresponding cost by the inner array position, as an integer.
  * Correct result is only possible by having the enum values in a consistent order with the integer array.
  * @see fighterClass
@@ -1638,143 +1221,6 @@ char *descStringFromSkill(skillType class)
 int costFromSkill(skillType class)
 {
     return skillscosts[class];
-}
-
-/**
- * Takes a integer and returns the corresponding perk name string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see perkClass
- * @see Perk
- * @param p The integer/perkClass.
- * @return String corresponding to the perk name.
- */
-char *nameStringFromPerk(int p)
-{
-
-    return perksnamestrings[p];
-}
-
-/**
- * Takes a integer and returns the corresponding perk desc string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see perkClass
- * @see Perk
- * @param p The integer/perkClass.
- * @return String corresponding to the perk desc.
- */
-char *descStringFromPerk(int p)
-{
-
-    return perksdescstrings[p];
-}
-
-/**
- * Takes a integer and returns the corresponding treasure desc string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see treasureClass
- * @param t The integer/treasureClass.
- * @return String corresponding to the treasure desc.
- */
-char *descStringFromTreasure(int t)
-{
-
-    return treasuredescstrings[t];
-}
-
-/**
- * Takes a integer and returns the corresponding chest desc string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see chestClass
- * @param c The integer/chestClass.
- * @return String corresponding to the chest desc.
- */
-char *descStringFromChest(int c)
-{
-    return chestdescstrings[c];
-}
-
-/**
- * Takes a integer and returns the corresponding FoeParty desc string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see foePartyClass
- * @param fp The integer/foePartyClass.
- * @return String corresponding to the FoeParty desc.
- */
-char *stringFromFoePartyClass(foePartyClass fp)
-{
-    return foepartystrings[fp];
-}
-
-/**
- * Takes a integer and returns the corresponding HLP_Region_Type string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see HLP_Region_Type
- * @param t The HLP type.
- * @return String corresponding to the HLP_Region_Type name.
- */
-char *stringFrom_HLP_Region_Type(HLP_Region_Type t)
-{
-    return hlp_regiontype_strings[t];
-}
-
-/**
- * Takes a integer and returns the corresponding room name string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see roomClass
- * @param r The integer/roomClass.
- * @return String corresponding to the room name.
- */
-char *stringFromRoom(roomClass r)
-{
-    return roomnamestrings[r];
-}
-
-/**
- * Takes a integer and returns the corresponding gamemode string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see Gamemode
- * @param g The integer/Gamemode.
- * @return String corresponding to the gamemode.
- */
-char *stringFromGamemode(Gamemode g)
-{
-    return gamemodenamestrings[g];
-}
-
-/**
- * Takes a integer and returns the corresponding floorClass string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see floorClass
- * @param fc The integer/floorClass.
- * @return String corresponding to the floorClass.
- */
-char *stringFromFloorclass(floorClass fc)
-{
-    return floornamestrings[fc];
-}
-
-/**
- * Takes a integer and returns the corresponding HLPD_KeyClass string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see HLPD_KeyClass
- * @param k The integer/HLPD_KeyClass.
- * @return String corresponding to the HLPD_KeyClass.
- */
-char *stringFrom_HLPD_KeyClass(HLPD_KeyClass k)
-{
-    return hlpd_keyclass_strings[k];
-}
-
-/**
- * Takes a integer and returns the corresponding HLPD_DirectionalKeys_Schema string by the inner array position.
- * Correct result is only possible by having the enum values in a consistent order with the string array.
- * @see HLPD_DirectionalKeys_Schema
- * @param dks The integer/HLPD_DirectionalKeys_Schema.
- * @return String corresponding to the HLPD_DirectionalKeys_Schema.
- */
-const char *stringFrom_HLPD_DirectionalKeys_Schema(int dks)
-{
-    return hlpd_directionalkeyschemas_strings[dks];
 }
 
 /**
@@ -5614,97 +5060,6 @@ bool check_seed(char buffer[PATH_SEED_BUFSIZE])
         }
     }
     return true;
-}
-
-void hlpd_reset_logfile(void)
-{
-#ifndef HELAPORDO_DEBUG_LOG
-#else
-    FILE *debug_file = NULL;
-    FILE *OPS_debug_file = NULL;
-    // Open log file if log flag is set and reset it
-    if (G_LOG_ON == 1) {
-        char path_to_debug_file[600];
-        char path_to_OPS_debug_file[600];
-        char static_path[500];
-        // Set static_path value to the correct static dir path
-        resolve_staticPath(static_path);
-
-        //Truncate "debug_log.txt"
-        sprintf(path_to_debug_file, "%s/%s", static_path, "debug_log.txt");
-        debug_file = fopen(path_to_debug_file, "w");
-        if (!debug_file) {
-#ifdef HELAPORDO_CURSES_BUILD
-            endwin();	//TODO: Can/should we check if we have to do this only in curses mode?
-#endif // HELAPORDO_CURSES_BUILD
-            fprintf(stderr,
-                    "[ERROR]    Can't open debug logfile (%s/debug_log.txt).\n",
-                    static_path);
-            exit(EXIT_FAILURE);
-        }
-        fprintf(debug_file, "[DEBUGLOG]    --New game--  \n");
-#ifdef HELAPORDO_CURSES_BUILD
-        if (NCURSES_VERSION_MAJOR < EXPECTED_NCURSES_VERSION_MAJOR
-            || (NCURSES_VERSION_MAJOR == EXPECTED_NCURSES_VERSION_MAJOR && NCURSES_VERSION_MINOR < EXPECTED_NCURSES_VERSION_MINOR)
-            || (NCURSES_VERSION_MAJOR == EXPECTED_NCURSES_VERSION_MAJOR && NCURSES_VERSION_MINOR == EXPECTED_NCURSES_VERSION_MINOR && NCURSES_VERSION_PATCH < EXPECTED_NCURSES_VERSION_PATCH)) {
-            fprintf(debug_file,
-                    "[WARN]    ncurses version is lower than expected {%s: %i.%i.%i} < {%i.%i.%i}\n",
-                    NCURSES_VERSION, NCURSES_VERSION_MAJOR,
-                    NCURSES_VERSION_MINOR, NCURSES_VERSION_PATCH,
-                    EXPECTED_NCURSES_VERSION_MAJOR,
-                    EXPECTED_NCURSES_VERSION_MINOR,
-                    EXPECTED_NCURSES_VERSION_PATCH);
-        }
-#else
-#ifndef HELAPORDO_RAYLIB_BUILD
-#error "HELAPORDO_CURSES_BUILD and HELAPORDO_RAYLIB_BUILD are both undefined"
-#else
-        if (RAYLIB_VERSION_MAJOR < EXPECTED_RAYLIB_VERSION_MAJOR
-            && RAYLIB_VERSION_MINOR < EXPECTED_RAYLIB_VERSION_MINOR
-            && RAYLIB_VERSION_PATCH < EXPECTED_RAYLIB_VERSION_PATCH) {
-            fprintf(debug_file,
-                    "[WARN]    raylib version is lower than expected {%s: %i.%i.%i} < {%i.%i.%i}\n",
-                    RAYLIB_VERSION, RAYLIB_VERSION_MAJOR,
-                    RAYLIB_VERSION_MINOR, RAYLIB_VERSION_PATCH,
-                    EXPECTED_RAYLIB_VERSION_MAJOR,
-                    EXPECTED_RAYLIB_VERSION_MINOR,
-                    EXPECTED_RAYLIB_VERSION_PATCH);
-        }
-#endif // HELAPORDO_RAYLIB_BUILD
-#endif // HELAPORDO_CURSES_BUILD
-        fprintf(debug_file, "[DEBUG]    --Default kls debug info:--  \n");
-        print_kls_2file(debug_file, default_kls);
-        fprintf(debug_file, "[DEBUG]    --Temporary kls debug info:--  \n");
-        print_kls_2file(debug_file, temporary_kls);
-        fprintf(debug_file,
-                "[DEBUG]    --Closing header for new game.--  \n");
-        fclose(debug_file);
-
-        //Lay debug info
-        log_tag("debug_log.txt", "[DEBUG]", "G_DEBUG_ON == (%i)",
-                G_DEBUG_ON);
-        log_tag("debug_log.txt", "[DEBUG]", "G_LOG_ON == (%i)", G_LOG_ON);
-        log_tag("debug_log.txt", "[DEBUG]", "small DEBUG FLAG ASSERTED");
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[Current position in default_kls] [pos: %li]\n",
-                kls_get_pos(default_kls));
-
-        //Truncate OPS_LOGFILE
-        sprintf(path_to_OPS_debug_file, "%s/%s", static_path, OPS_LOGFILE);
-        OPS_debug_file = fopen(path_to_OPS_debug_file, "w");
-        if (!OPS_debug_file) {
-#ifdef HELAPORDO_CURSES_BUILD
-            endwin();	//TODO: Can/should we check if we have to do this only in curses mode?
-#endif // HELAPORDO_CURSES_BUILD
-            fprintf(stderr, "[ERROR]    Can't open OPS logfile (%s/%s).\n",
-                    static_path, OPS_LOGFILE);
-            exit(EXIT_FAILURE);
-        }
-        fprintf(OPS_debug_file, "[OPLOG]    --New game--  \n");
-        fclose(OPS_debug_file);
-        log_tag("debug_log.txt", "[DEBUG]", "Truncated [%s]", OPS_LOGFILE);
-    }
-#endif
 }
 
 void hlpd_use_forced_flags(const char* whoami)
